+++ Lab6.cbp
@@ -31,6 +31,8 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="Samochody.cpp" />
+		<Unit filename="Samochody.h" />
 		<Unit filename="main.cpp" />
 		<Extensions>
 			<code_completion />
