+++ main.cpp
@@ -1,57 +1,9 @@
 #include <iostream>
 #include <string>
+#include "Samochody.h"
 
 using namespace std;
 
-class Samochody
-{
-public:
-    void start()
-    {
-    cout<<"\nWitam. Wywolane metoda start. \n"<<endl;
-    }
-    Samochody();
-    Samochody(string marka,string model,int rok, float pojemnosc);
-    ~Samochody();
-    void save(string marka,string model,int rok, float pojemnosc);
-    void read(string marka,string model,int rok, float pojemnosc);
-private:
-    string marka,model;
-    int rok;
-	float pojemnosc;
-
-};
-
-    Samochody::Samochody()
-    {
-
-    }
-    Samochody::~Samochody()
-    {
-
-    }
-    Samochody::Samochody(string marka,string model,int rok, float pojemnosc)
-    {
-    marka="Renault";
-    model="Clio";
-    rok='2001';
-    pojemnosc='1149';
-    }
-
-void Samochody::save(string mar,string mod,int ro,float poj)
-    {
-        marka=mar;
-        model=mod;
-        rok = ro;
-        pojemnosc = poj;
-    }
-void Samochody::read(string marka,string model,int rok,float pojemnosc)
-    {
-        cout<<"\nMarka: "<<marka<<endl;
-        cout<<"Model: "<<model<<endl;
-        cout<<"Rok: "<<rok<<endl;
-        cout<<"Pojemnosc: "<<pojemnosc<<endl;
-    }
 
 int main()
 {
